# baas-swagger-v7.3.yaml
swagger: "2.0"
info:
  description: "This is a basic documentation of the Blockchain as a Service (BaaS), platform that will be integrated in ioTFeds by CERTH/ITI. The baasic endpoints for user and federation magement are described here. The documentation will be adjusted and extended as the platform is integrated."
  version: "1.0.0"
  title: "On-chain information in IoTFeds"
# # host: "petstore.swagger.io"
# # basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /removePlatformResources:
      delete:
        tags:
        - "user"
        summary: "Removes all resources from a platform"
        description: "Removes all resources from a platform"
        operationId: "removePlatformresources"
        produces:
        - "message"
        parameters:
        - name: "platform_id"
          in: "query"
          description: "The platform's id"
          required: true
          type: "string"
        responses:
          "200":
            description: "successful operation"
          "400":
            description: "Deletion failed"
          "404":
            description: "Platform resources not found"
...